pyproject.toml
[project]
name = "hostamar-local"
version = "0.1.0"
description = "Hostamar Enterprise Platform (Local)"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "pydantic",
    "sqlalchemy",
    "boto3",
    "langgraph",
    "langchain-openai",
    "langchain-core",
    "python-dotenv",
    "playwright",
    "redis",
    "jinja2",
    "pyyaml",
    "pytesseract",
    "pillow",
    "openpyxl",
    "pandas",
    "num2words",
    "requests",
    "argon2-cffi"
]
